requirements.txt
Django==5.1.3
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
drf-yasg==1.21.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
